eng84-airport-project/requirements.txt
beautifulsoup4==4.9.3
Django==3.1.7
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==2.3.1
importlib-metadata==2.1.1
pytz==2021.1
soupsieve==2.2.1
sqlparse==0.4.1
zipp==3.4.1
